refactor: 패키지 구조 변경 및 불필요한 import 문 제거
- validation 관련 클래스를 application 모듈로 이동
- infrastructure 모듈 내 DB 패키지 재구성
- import 및 package 문 정정
