+++ movie-application/src/main/java/com/example/application/service/MovieSearchRateLimiterService.java
@@ -3,7 +3,6 @@ package com.example.application.service;
 import com.google.common.util.concurrent.RateLimiter;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.TimeUnit;
 import org.springframework.stereotype.Service;
 
 @Service
