+++ movie-infrastructure/src/main/java/com/example/infrastructure/config/CacheConfig.java
@@ -1,4 +1,4 @@
-package com.example.infrastructure.db.config;
+package com.example.infrastructure.config;
 
 import java.time.Duration;
 import lombok.RequiredArgsConstructor;
