+++ movie-infrastructure/src/main/java/com/example/infrastructure/config/JpaConfig.java
@@ -1,4 +1,4 @@
-package com.example.infrastructure.db.config;
+package com.example.infrastructure.config;
 
 import java.util.Optional;
 import org.springframework.context.annotation.Bean;
