+++ movie-infrastructure/src/main/java/com/example/infrastructure/config/RedissonConfig.java
@@ -1,4 +1,4 @@
-package com.example.infrastructure.db.config;
+package com.example.infrastructure.config;
 
 import org.redisson.Redisson;
 import org.redisson.api.RedissonClient;
