+++ movie-infrastructure/src/main/java/com/example/infrastructure/db/persistence/MemberJpaRepositoryAdapter.java
@@ -1,4 +1,4 @@
-package com.example.infrastructure.persistence;
+package com.example.infrastructure.db.persistence;
 
 import com.example.application.port.out.MemberRepositoryPort;
 import com.example.domain.model.entity.Member;
