+++ movie-infrastructure/src/main/java/com/example/infrastructure/db/persistence/TheaterJpaRepositoryAdapter.java
@@ -1,4 +1,4 @@
-package com.example.infrastructure.persistence;
+package com.example.infrastructure.db.persistence;
 
 import com.example.application.port.out.TheaterRepositoryPort;
 import com.example.domain.model.entity.Theater;
