+++ movie-infrastructure/src/test/java/com/example/infrastructure/ReservationConcurrencyTest.java
@@ -6,13 +6,10 @@ import com.example.application.dto.request.ReservationRequestDto;
 import com.example.application.port.in.ReservationServicePort;
 import com.example.application.port.out.ReservationRepositoryPort;
 import com.example.application.port.out.ScreeningSeatRepositoryPort;
-import com.example.domain.exception.CustomException;
-import com.example.domain.exception.ErrorCode;
 import java.util.List;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
-import java.util.concurrent.atomic.AtomicInteger;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
